backend/requirements.txt
Django==4.2.20
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==25.1
psycopg2-binary==2.9.9
newrelic==10.9.0 